code/deployment/requirements.txt
fastapi
streamlit
uvicorn
Pillow
numpy
starlette
onnxruntime
python-multipart